+++ kube/src/lib.rs
@@ -9,22 +9,23 @@
 //! and then watch for it to become available
 //!
 //! ```rust,no_run
+//! use futures::{StreamExt, TryStreamExt};
+//! use kube::api::{Api, ListParams, PostParams, WatchEvent};
+//! use kube::Client;
+//! use kube::runtime::Informer;
 //! use k8s_openapi::api::core::v1::Pod;
-//! use tokio::stream::StreamExt as _;
-//! use futures_util::stream::StreamExt as _;
 //!
-//! async {
+//! #[tokio::main]
+//! async fn main() -> Result<(), kube::Error> {
 //!     // Read the environment to find config for kube client.
 //!     // Note that this tries an in-cluster configuration first,
 //!     // then falls back on a kubeconfig file.
-//!     let kube_client = kube::Client::try_default()
-//!        .await
-//!        .expect("kubeconfig failed to load");
-//!     
+//!     let kube_client = Client::try_default().await?;
+//!
 //!     // Get a strongly typed handle to the Kubernetes API for interacting
 //!     // with pods in the "default" namespace.
-//!     let pods: kube::Api<Pod> = kube::Api::namespaced(kube_client.clone(), "default");
-//!     
+//!     let pods: Api<Pod> = Api::namespaced(kube_client, "default");
+//!
 //!     // Create a pod from JSON
 //!     let pod = serde_json::from_value(serde_json::json!({
 //!         "apiVersion": "v1",
@@ -40,26 +41,23 @@
 //!                 },
 //!             ],
 //!         }
-//!     })).unwrap();
+//!     }))?;
 //!
 //!     // Create the pod
-//!     let pod = pods.create(&kube::api::PostParams::default(), &pod).await.unwrap();
+//!     let pod = pods.create(&PostParams::default(), &pod).await?;
 //!
 //!     // Create an informer for watching events about
-//!     let informer: kube::runtime::Informer<Pod> = kube::runtime::Informer::new(
-//!         kube_client,
-//!         kube::api::ListParams::default()
+//!     let informer = Informer::new(pods,
+//!         ListParams::default()
 //!             .fields("metadata.name=my-container")
 //!             .timeout(10),
-//!         kube::Resource::namespaced::<Pod>("default"),
 //!     );
 //!
 //!     // Get an event stream from the informer
-//!     let mut events_stream = informer.poll().await.unwrap().boxed();
-//!     
+//!     let mut events_stream = informer.poll().await?.boxed();
+//!
 //!     // Keep getting events from the events stream
-//!     while let Some(event) = events_stream.try_next().await.unwrap() {
-//!         use kube::api::WatchEvent;
+//!     while let Some(event) = events_stream.try_next().await? {
 //!         match event {
 //!             WatchEvent::Modified(e) if e.status.as_ref().unwrap().phase.as_ref().unwrap() == "Running" => {
 //!                 println!("It's running!");
@@ -70,7 +68,8 @@
 //!             _ => {}
 //!         }
 //!     }
-//! };
+//!     Ok(())
+//! }
 //! ```
 
 #![deny(missing_docs)]